Parse NA queenstown xblock (#700)
* Fix: #696

* Update Maple2.File.Parser to 2.0.6

* Update ms2-map-metadata-hash

+++ MapleServer2/Managers/FieldManager.cs
@@ -28,7 +28,7 @@ public class FieldManager
 
     public readonly int MapId;
     public readonly long InstanceId;
-    public readonly short Capacity;
+    public readonly int Capacity;
     public readonly bool IsTutorialMap;
     public readonly FieldState State = new();
     public readonly CoordS[] BoundingBox;
